requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-debug-toolbar
pyjwt